# Speed up lookup of possible types for Fragments